requirements.txt
python-dotenv
qdrant_client
sentence-transformers
numpy
fastembed
bs4
colorama